Refactor js functions into separate files
